+++ plot/viewer/src/canvases/SketchCanvas.tsx
@@ -12,6 +12,7 @@ import type {
   SketchNode as DocNode,
 } from "../types";
 import { SketchContextMenu } from "./SketchContextMenu";
+import { EDGE_TYPES } from "./edges/registry";
 import { NODE_RENDERERS } from "./nodes/registry";
 import { SketchToolbar } from "./SketchToolbar";
 import { useSketchClipboard } from "./useSketchClipboard";
@@ -311,6 +312,7 @@ function SketchCanvasInner({
         nodes={nodes}
         edges={edges}
         nodeTypes={NODE_TYPES}
+        edgeTypes={EDGE_TYPES}
         nodesConnectable
         nodesDraggable
         elementsSelectable